Readers

Axl reader

It is possible to read Axl files and get the list of geometric objects contains in these files, using the function axlread.

Axl.axlread — Function

Read an axl file and ouput the list of objets:

axlread("file.axl")

The function uses the table Axl.Reader::Dict{String,Function} to determine how to read the elements in the file. The functions needed to read an object of a given type "key" are defined as follows:

Reader["key"] = function(obj::EzXML.Node) ... end

If the element obj has a key type then the reader Reader[obj["type"]] is used, otherwise the reader Reader[nodename(obj)] is used.

Here is an example of objects saved in the file tmp.axl with the command @axlview and read with the function axlread.

s = sphere(point(0.,1.,0.), 0.5, color=Color(255,0,0))
c = cube(point(0.,0.,0.), 0.5, field=[Float64(i) for i in 1:8])
@axlview s,c
l = axlread("tmp.axl")

The list l contains a copy of the sphere s and of the mesh c.

Obj reader

Axl.objread — Function

Read an obj file and ouput a mesh.

objread("file.obj")