Viewer

The objects in SemiAlgebraicTypes can be vizualized, using Axl. They are printed in a file tmp.axl and Axl is then run on this view file.

@axl start

Initialize the view file.

@axl m

Add the object m to the viewfile.

@axl view

Open Axl for visualization of the data in the view file.

@axlview

This is equivalent to @axl view.

@axlview m

Save the object m in the viewfile and visualize it

Example

using Axl
@axl start
A = point(0.,0.5,0.)
B = point(0.,1.5,0.)
C = point(0.,3.5,0.)

@axl c0 = cylinder(A,B,0.2, color=Color(255,0,0))
@axl c1 = cone(C,B,0.7, color=Color(0,255,0))
@axl view

@axlview c0, sphere(A,0.5)

Functions

Axl.@axl — Macro

Initiliaze an Axl view, or add objects to the view or visualize using Axl viewer.

@axl start
@axl O = point(0.,0.,0.)
A = point(1.,0.,0.); S = sphere(A, 0.5, color=Color(255,0,0)); @axl A,S
@axl view

Axl.@axlview — Macro

Display a geometric object with Axl. If no argument is given, the last view is used if it exists.

@axlview A = point(1.,1.,1.)
A=point(1.,1.,0.); S=sphere(point(0.,0.,0.),0.5, color=color(255,0,0)); @axlview A,S
@axlview

Axl.objprint — Function

Print a mesh in an obj file.

objprint(m, "file.obj")